Finished project, added sticky search bar and cleaned up the code a bit, added more comments to the functionality of react in relation to the app
