+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App'; //Changing CardList to App as we're moving to
+import App from './containers/App'; //Changing CardList to App as we're moving to
                          //the final stages of the app
 import * as serviceWorker from './serviceWorker';
 import 'tachyons';
